feat(phy): OFDM symbol path, preamble, passband converters, frame transmitter (P1-4)
- phy/ofdm.py: DC-centred carrier map with comb pilots on both edges, one
  Zadoff-Chu pilot sequence (full pilot symbols ≈ 3.5 dB PAPR), windowed
  overlap-add symbols (8-sample raised-cosine taper, effective CP 5 ms),
  demodulator with mid-CP FFT window and phase-ramp compensation —
  noiseless EVM is 0 (legacy: 18.7 %)
- phy/preamble.py: two Schmidl-Cox symbols (ZC on even carriers) and a
  unique word from a prime-length (53) ZC cyclically extended to 57
  carriers whose root encodes the frame header (14 data modes + control)
- phy/passband.py: streaming 8 kHz complex baseband <-> 48 kHz real audio
  at 1 500 Hz centre with band-limiting and interpolation FIRs (70 dB),
  block-invariant, integer group delay at baseband
- phy/tx.py: frame assembler (preamble + pilot/data symbols) and burst/audio
  helpers; preamble and data symbols within 1 dB of each other (TX-envelope
  requirement from COMMUNITY-CONCERNS #8)
- fec/nr_ldpc.py: decoder uses precomputed gather/scatter tables (3x faster
  single-codeword decoding)
- tests: 99 % power inside ±1 180 Hz, < −40 dBc beyond 300 Hz past the edge
  carriers, frame survives the audio path at < −40 dB EVM

+++ tools/bench_ldpc.py
@@ -42,7 +42,7 @@ def run_point(
 ) -> dict[str, float | int | str]:
     z = select_lifting_size(2, info_len)
     code = nr_ldpc_code(2, z)
-    e = int(round(info_len / rate))
+    e = round(info_len / rate)
     rm = RateMatcher(code, info_len, e, rv=0)
     bpsk = constellation(Modulation.BPSK)
     noise_var = 1.0 / 10 ** (es_n0_db / 10)
